pyproject.toml
[tool.poetry]
name = "enfify"
version = "0.1.0"
description = "Detection of audio manipulation by using electrical network frequency."
authors = ["Licates <136702730+Licates@users.noreply.github.com>, Dacasil <https://github.com/Dacasil>"]
readme = "README.md"

[tool.poetry.scripts]
enfify = "enfify.main:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12.5"
loguru = "^0.7.2"
numpy = "^2.1.1"
tqdm = "^4.66.5"
python-dotenv = "^1.0.1"
scipy = "^1.14.1"
matplotlib = "^3.9.2"
ffmpeg-python = "^0.2.0"
pyyaml = "^6.0.2"
requests = "^2.32.3"
torch = {version = "^2.4.1+cpu", source = "pytorch-cpu"}
scikit-learn = "1.5.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
torchsummary = "^1.5.1"


[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.ruff.lint.isort]
known_first_party = ["enfify"]
force_sort_within_sections = true
